Extract rules to Given/When/Then filenames in router/
- Makes existing Rules more visible/discoverable/reusable to devs extending the application
- Reduces vertical scroll in index.tsx
