+++ src/pages/manage-my-registrations/delete/index.tsx
@@ -1,4 +1,4 @@
-import { GetServerSideProps, GetServerSidePropsResult } from "next";
+import { GetServerSideProps } from "next";
 import React, { FunctionComponent } from "react";
 import { BeaconsForm } from "../../../components/BeaconsForm";
 import { Button, LinkButton } from "../../../components/Button";
@@ -8,25 +8,18 @@ import { RadioList, RadioListItem } from "../../../components/RadioList";
 import { SummaryList, SummaryListItem } from "../../../components/SummaryList";
 import { Beacon } from "../../../entities/Beacon";
 import { ReasonsForDeletingARegistration } from "../../../entities/ReasonsForDeletingARegistration";
-import { showCookieBanner } from "../../../lib/cookies";
 import { FieldManager } from "../../../lib/form/FieldManager";
 import { FormJSON, FormManager } from "../../../lib/form/FormManager";
-import {
-  isInvalid,
-  isValid,
-  withErrorMessages,
-  withoutErrorMessages,
-} from "../../../lib/form/lib";
 import { Validators } from "../../../lib/form/Validators";
-import { FormManagerFactory } from "../../../lib/handlePageRequest";
 import { BeaconsGetServerSidePropsContext } from "../../../lib/middleware/BeaconsGetServerSidePropsContext";
 import { withContainer } from "../../../lib/middleware/withContainer";
 import { withSession } from "../../../lib/middleware/withSession";
-import { redirectUserTo } from "../../../lib/redirectUserTo";
 import { PageURLs } from "../../../lib/urls";
 import { prettyUseName } from "../../../lib/writingStyle";
 import { BeaconsPageRouter } from "../../../router/BeaconsPageRouter";
-import { Rule } from "../../../router/rules/Rule";
+import { GivenUserIsDeletingARegistration_WhenUserDoesNotProvideAReason_ThenShowErrorMessage } from "../../../router/rules/GivenUserIsDeletingARegistration_WhenUserDoesNotProvideAReason_ThenShowErrorMessage";
+import { GivenUserIsDeletingARegistration_WhenUserProvidesAReason_ThenDeleteTheRegistration } from "../../../router/rules/GivenUserIsDeletingARegistration_WhenUserProvidesAReason_ThenDeleteTheRegistration";
+import { GivenUserIsDeletingARegistration_WhenUserViewsPage_ThenDisplayPage } from "../../../router/rules/GivenUserIsDeletingARegistration_WhenUserViewsPage_ThenDisplayPage";
 
 export interface DeleteRegistrationProps {
   form: FormJSON;
@@ -35,7 +28,7 @@ export interface DeleteRegistrationProps {
   previousPageURL?: PageURLs;
 }
 
-interface DeleteRegistrationForm {
+export interface DeleteRegistrationForm {
   reasonForDeletion: ReasonsForDeletingARegistration;
 }
 
@@ -159,160 +152,21 @@ export const DeleteRegistration: FunctionComponent<DeleteRegistrationProps> = ({
 export const getServerSideProps: GetServerSideProps = withSession(
   withContainer(async (context: BeaconsGetServerSidePropsContext) => {
     return await new BeaconsPageRouter([
-      new WhenUserViewsDeleteRegistrationPage_ThenDisplayPage(
+      new GivenUserIsDeletingARegistration_WhenUserViewsPage_ThenDisplayPage(
         context,
         validationRules
       ),
-      new GivenUserHasNotSelectedAReason_WhenUserTriesToDeleteARegistration_ThenShowErrorMessage(
+      new GivenUserIsDeletingARegistration_WhenUserDoesNotProvideAReason_ThenShowErrorMessage(
         context,
         validationRules
       ),
-      new GivenUserHasSelectedAReason_WhenUserTriesToDeleteARegistration_ThenDeleteTheRegistration(
+      new GivenUserIsDeletingARegistration_WhenUserProvidesAReason_ThenDeleteTheRegistration(
         context,
         validationRules
       ),
     ]).execute();
   })
 );
-
-class WhenUserViewsDeleteRegistrationPage_ThenDisplayPage implements Rule {
-  private context: BeaconsGetServerSidePropsContext;
-  private validationRules: FormManagerFactory;
-
-  constructor(
-    context: BeaconsGetServerSidePropsContext,
-    validationRules: FormManagerFactory
-  ) {
-    this.context = context;
-    this.validationRules = validationRules;
-  }
-
-  async condition(): Promise<boolean> {
-    return this.context.req.method === "GET";
-  }
-
-  async action(): Promise<GetServerSidePropsResult<DeleteRegistrationProps>> {
-    const { getBeaconsByAccountHolderId, getOrCreateAccountHolder } =
-      this.context.container;
-
-    const accountHolder = await getOrCreateAccountHolder(this.context.session);
-    const registrationToBeDeleted = (
-      await getBeaconsByAccountHolderId(accountHolder.id)
-    ).find((beacon) => beacon.id == this.context.query.id);
-
-    return {
-      props: {
-        form: withoutErrorMessages({}, this.validationRules),
-        previousPageURL: PageURLs.accountHome,
-        beacon: registrationToBeDeleted,
-        showCookieBanner: true,
-      },
-    };
-  }
-}
-
-class GivenUserHasNotSelectedAReason_WhenUserTriesToDeleteARegistration_ThenShowErrorMessage {
-  private readonly context: BeaconsGetServerSidePropsContext;
-  private readonly validationRules: FormManagerFactory;
-
-  constructor(
-    context: BeaconsGetServerSidePropsContext,
-    validationRules: FormManagerFactory
-  ) {
-    this.context = context;
-    this.validationRules = validationRules;
-  }
-
-  async condition(): Promise<boolean> {
-    return (
-      this.context.req.method === "POST" &&
-      isInvalid(await this.form(), this.validationRules)
-    );
-  }
-
-  async action(): Promise<GetServerSidePropsResult<DeleteRegistrationProps>> {
-    const { getBeaconsByAccountHolderId, getOrCreateAccountHolder } =
-      this.context.container;
-
-    const accountHolder = await getOrCreateAccountHolder(this.context.session);
-    const registrationToBeDeleted = (
-      await getBeaconsByAccountHolderId(accountHolder.id)
-    ).find((beacon) => beacon.id == this.context.query.id);
-
-    return {
-      props: {
-        beacon: registrationToBeDeleted,
-        form: await withErrorMessages(this.form(), this.validationRules),
-        showCookieBanner: showCookieBanner(this.context),
-      },
-    };
-  }
-
-  private async form(): Promise<FormJSON> {
-    return await this.context.container.parseFormDataAs<FormJSON>(
-      this.context.req
-    );
-  }
-}
-
-class GivenUserHasSelectedAReason_WhenUserTriesToDeleteARegistration_ThenDeleteTheRegistration {
-  private readonly context: BeaconsGetServerSidePropsContext;
-  private readonly validationRules: FormManagerFactory;
-
-  constructor(
-    context: BeaconsGetServerSidePropsContext,
-    validationRules: FormManagerFactory
-  ) {
-    this.context = context;
-    this.validationRules = validationRules;
-  }
-
-  public async condition(): Promise<boolean> {
-    return (
-      this.context.req.method === "POST" &&
-      isValid(await this.form(), this.validationRules)
-    );
-  }
-
-  public async action(): Promise<
-    GetServerSidePropsResult<DeleteRegistrationProps>
-  > {
-    const { deleteBeacon } = this.context.container;
-
-    const success: boolean = (
-      await deleteBeacon(
-        await this.reasonForDeletion(),
-        this.registrationId(),
-        await this.accountHolderId()
-      )
-    ).success;
-
-    if (success) return redirectUserTo(PageURLs.deleteRegistrationSuccess);
-    else redirectUserTo(PageURLs.deleteRegistrationFailure);
-  }
-
-  private async form(): Promise<DeleteRegistrationForm> {
-    return await this.context.container.parseFormDataAs<DeleteRegistrationForm>(
-      this.context.req
-    );
-  }
-
-  private registrationId(): string {
-    return this.context.query.id as string;
-  }
-
-  private async accountHolderId(): Promise<string> {
-    return (
-      await this.context.container.getOrCreateAccountHolder(
-        this.context.session
-      )
-    ).id;
-  }
-
-  private async reasonForDeletion(): Promise<ReasonsForDeletingARegistration> {
-    return (await this.form()).reasonForDeletion;
-  }
-}
 
 const validationRules = ({
   reasonForDeletion,
